+++ src/pages/comparison.tsx
@@ -1,3 +1,0 @@
-import { Comparison } from 'src/views/Comparison';
-
-export default Comparison;