+++ src/services/pizza/types.ts
@@ -9,9 +9,10 @@ export type City = typeof supportedCities[number];
 
 type Title = `${Lang}_title`;
 type Description = `${Lang}_description`;
+type Company = `${Lang}_company`;
 
 export type HistoryOfChangesWatchKey = Title | Description | 'image' | 'price' | 'weight';
-export type HistoryOfChangesValue = string | number | undefined;
+export type HistoryOfChangesValue = string | number | undefined | null;
 
 export interface Change {
   key: HistoryOfChangesWatchKey;
@@ -35,4 +36,6 @@ interface BasePizza {
 
 export type Pizza = BasePizza & {
   [key in Title | Description]: string;
+} & {
+  [key in Company]: string;
 };