+++ src/views/Pizza/style.ts
@@ -83,3 +83,13 @@ export const HistoryList = styled.ul`
     color: ${color('bold')};
   }
 `;
+
+export const Company = styled.a`
+  font-weight: ${fontWeight('bold')};
+`;
+
+export const ChangedValue = styled.span<{ isValue?: boolean }>`
+  font-weight: ${fontWeight('bold')};
+
+  ${(p) => p.isValue && `color: ${color('bold')(p)};`}
+`;